+++ Changes.md
@@ -1,6 +1,7 @@
 HEAD
 -----------
 - Handle more ZK exceptions as candidates for reconnecting the client on error.
+- Add safety check to actively purge redis clients if a RedisFailover::Client hasn't heard from the Node Manager in a certain time window.
 
 0.5.2
 -----------